refactor(api): break down preview usecase into services
