+++ apps/api/src/app/workflows-v2/usecases/preview/utils/variable-helpers.ts
@@ -1,0 +1,8 @@
+import _ from 'lodash';
+
+/**
+ * Replaces all occurrences of a search string with a replacement string.
+ */
+export function replaceAll(text: string, searchValue: string, replaceValue: string): string {
+  return _.replace(text, new RegExp(_.escapeRegExp(searchValue), 'g'), replaceValue);
+}
